==> Notebook/1. Data Cleaning (Basic).ipynb <==
### Data Exploration and Cleanup


import pandas as pd
import numpy as np

!ls

#### Reading DataSet

df = pd.read_csv('../Data/raw_data.csv', low_memory = False, compression='gzip')
len(df)

df.head()

### A Customer_ID `number` at a particular timestamp can only have one entry
### Removing Duplicate Entries ['ts','number']

df[df.duplicated(subset=['ts','number'],keep=False)]

### There are 113540 Duplicate Entries
#### We have 8315498 Unique timestamp, customer_id rows. 

## Keeping first occurence
df.drop_duplicates(subset=['ts','number'], inplace = True, keep = 'last')

df.reset_index(inplace = True, drop = True)

# Info of Dataset
df.info()

#Count missing values
df.isnull().values.sum()

df['number'] = pd.to_numeric(df['number'], errors = 'coerce')

#Count missing values
np.count_nonzero(df.isnull().values)

#### There are 116 NaN rows, dropping NaN rows.

df.dropna(inplace = True)
len(df)

df['number'] = pd.to_numeric(df['number'], errors = 'coerce', downcast='integer')
df['ts'] = pd.to_datetime(df['ts'])

# Info of Dataset
df.info()

### Breaking Time to Features

df['hour'] = df['ts'].dt.hour
df['mins'] = df['ts'].dt.minute
df['day'] = df['ts'].dt.day
df['month'] = df['ts'].dt.month
df['year'] = df['ts'].dt.year
df['dayofweek'] = df['ts'].dt.dayofweek

df.to_csv('../Data/data_checkpoint/preprocessed_1.csv',index = False, compression = 'gzip')

df
